Split up code to be more manageable
