+++ crit_app/assets/base.css
@@ -1,0 +1,19 @@
+/* Headers */
+h1, h2, h3, h4, h5, h6 {
+    font-weight: 600 !important;
+}
+
+/* Link styling */
+a {
+    text-decoration: none !important;
+    transition: color 0.2s ease !important;
+}
+
+a:hover {
+    color: #008062 !important;
+}
+
+/* Improve spacing */
+.mb-3 {
+    margin-bottom: 1.5rem !important;
+}
