+++ crit_app/assets/fonts.css
@@ -1,0 +1,4 @@
+@font-face {
+    font-family: "job-icons";
+    src: url("/assets/job-icons.ttf");
+}
