+++ crit_app/callbacks/__init__.py
@@ -1,0 +1,1 @@
+"""Callback functions for the How Bad Was My Crit In XIV application."""
